+++ src/characteristic/positionnable/Positionnable.java
@@ -5,5 +5,5 @@ import javafx.geometry.Point3D;
 public interface Positionnable extends Positionnable2D{
 	public void setPosition(Point3D position);
 	public Point3D getPosition();
-	public double distanceFrom(Positionnable p);
+	public float distanceFrom(Positionnable p);
 }
