+++ src/characteristic/positionnable/Reachable.java
@@ -3,6 +3,6 @@ package characteristic.positionnable;
 import characteristic.MessageReceiver;
 
 public interface Reachable extends Positionnable, MessageReceiver{
-    double computeReachableRadius();
-    double getReachableRadius();
+    float computeReachableRadius();
+    float getReachableRadius();
 }
